Trying to increase the numerical accuracy for virial integration as its sensitive for the constant volume stepping

+++ src/dynamo/dynamo/interactions/potentials/lennard_jones.hpp
@@ -72,7 +72,7 @@ namespace dynamo {
 
     virtual void outputXML(magnet::xml::XmlStream&) const;
 
-    double B2func(double) const;
+    long double B2func(long double) const;
 
     double _sigma;
     double _epsilon;
